import-maps/no-referencing-script-integrity-valid.html

<!DOCTYPE html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
let log;
</script>
<script type="importmap" onload="t.assert_unreached('onload')" onerror="t.done()">
{
  "integrity": {
    "./resources/log.js?pipe=sub&name=NoReferencingScriptValidCheck": "sha384-5eRmXQSBE6H5ENdymdZxcyiIfJL1dxtH8p+hOelZY7Jzk+gt0gYyemrGY0cEaThF"
  }
}
</script>
</head>
<body>
<script>
log = [];
let promiseResolve;
let promiseReject;
let promise = new Promise((resolve, reject) => {
  promiseResolve = resolve;
  promiseReject = reject;
});
</script>
<img src="/images/green.png?2"
  onload="import('./resources/log.js?pipe=sub&name=NoReferencingScriptValidCheck').then(promiseResolve).catch(promiseReject)">
<script>
promise_test(async () => {
  try {
    await promise;
  } catch {
    assert_unreached();
  }
  assert_equals(log.length, 1);
  assert_equals(log[0], "log:NoReferencingScriptValidCheck");
}, "Script was loaded as its valid integrity check passed");
</script>
</body>
</html>
